Errorhandling in frontend
