+++ FindYourFriends/App.js
@@ -8,16 +8,59 @@ import ChatScreen from "./screens/ChatScreen";
 import { ApolloClient } from "apollo-client";
 import { InMemoryCache } from "apollo-cache-inmemory";
 import { ApolloProvider } from "@apollo/react-hooks";
-import { authLink, errorLink, httpLink } from "./utils/links";
+import authLink from "./utils/links";
 import { createHttpLink } from "apollo-link-http";
+import { onError } from "apollo-link-error";
+import MyAlert from "./utils/MyAlert";
 
 export default function App() {
   // USE SCREENS LIKE THIS
   const [test, setTest] = useState(true);
   const [signedIn, setSignedIn] = useState(false);
-  const [error, setError] = useState(null);
+  const [error, setError] = useState({
+    message: null,
+    title: "An Error Occurred",
+  });
 
   const backendUri = "http://c4c25ff2.ngrok.io";
+  /**
+   * Put logic here, on how to handle errors.
+   * This handles all GraphQL errors.
+   * It was not my intention to put this here. I wanted it to be in utils/links.
+   * But I couldn't figure out how to get setError to utils/links.
+   */
+  const errorLink = onError(({ graphQLErrors, networkError }) => {
+    if (graphQLErrors) {
+      graphQLErrors.map((err) => {
+        const { message, locations, path } = err;
+        console.log(
+          `[GraphQL error]: 
+          Message: ${message}, 
+          Location: ${locations}, 
+          Path: ${path}, 
+          \n\nFull Error: ${err}`
+        );
+        if (err.extensions.code == "UNAUTHENTICATED") {
+          // Unauthenticated Error from backend.
+          // Add logic here like "If not authenticated, send to login-page"
+          // And set errorMessage to user via setError,
+          // that they tried something that required login, but they weren't logged in
+          setError({ message, title: "Unauthenticated." });
+        } else if (err.extensions.code == "FORBIDDEN") {
+          // ForbiddenError from backend.
+          // Should probably also send to login.
+          setError({
+            message: `You have to be logged in to do that.\n${message}`,
+            title: "Unauthorized action.",
+          });
+        } else {
+          // Open Alert box with message.
+          setError({ ...error, message });
+        }
+      });
+    }
+    if (networkError) console.log(`[Network error]: ${networkError}`);
+  });
   // the URI key is a string endpoint or function resolving to an endpoint -- will default to "/graphql" if not specified
   const httpLink = createHttpLink({ uri: backendUri + "/graphql" });
   /** httpLink and cache are requirements as of Apollo 2 https://www.apollographql.com/docs/react/api/apollo-client/#required-fields
@@ -50,7 +93,7 @@ export default function App() {
           text: buttonText,
           onPress: () => {
             console.log(JSON.stringify({ error }), null, 4);
-            setError(null);
+            setError({ message: null, title: "An Error Occurred" });
             console.log(`${buttonText} pressed on Error Alert.`);
           },
         },
@@ -75,7 +118,7 @@ export default function App() {
   return (
     <ApolloProvider client={client}>
       <View style={styles.screen}>
-        {error && MyAlert(error)}
+        {error.message && MyAlert(error.message, error.title)}
         <Header title="Find Your Friends" />
         {content}
       </View>
