+++ FindYourFriends/utils/links.js
@@ -1,31 +1,5 @@
-import { onError } from "apollo-link-error";
 import { setContext } from "apollo-link-context";
 import * as SecureStore from "expo-secure-store";
-
-/**
- * Put logic here, on how to handle errors.
- */
-const errorLink = onError(({ graphQLErrors, networkError }) => {
-  if (graphQLErrors) {
-    graphQLErrors.map((err) => {
-      const { message, locations, path } = err;
-      console.log(
-        `[GraphQL error]: 
-          Message: ${message}, 
-          Location: ${locations}, 
-          Path: ${path}, 
-          \n\nFull Error: ${err}`
-      );
-
-      if (err.extensions.code == "UNAUTHENTICATED") {
-        // Add logic here like "If not authenticated, send to login-page"
-        // And set errorMessage to user somewhere,
-        // that they tried something that required login, but they weren't logged in
-      }
-    });
-  }
-  if (networkError) console.log(`[Network error]: ${networkError}`);
-});
 
 /**
 The setContext function takes a function that returns either an object or a promise that returns an object to set the new context of a request.
@@ -46,4 +20,4 @@ const authLink = setContext(async (request, previousContext) => {
   };
 });
 
-export { authLink, errorLink };
+export default authLink;
